worker/requirements.txt
celery>=5.4.0
redis>=5.0.0
pydantic-settings>=2.2.0
SQLAlchemy>=2.0.28
asyncpg>=0.29.0
boto3>=1.34.0
ultralytics>=8.3.0
pyyaml>=6.0.1
Pillow>=10.0.0
albumentations>=1.4.0
opencv-python-headless>=4.9.0
